api/requirements.txt
gcsfs==2023.6.0
google-cloud-storage==2.9.0
google-cloud-run==0.9.0
google-cloud-firestore==2.11.1
python-dotenv==1.0.0
dbt-bigquery==1.5.3
dbt-core==1.5.2
msgpack==1.0.5
pydantic==2.0.3
pydantic_core==2.3.0
fastapi==0.100.0
uvicorn==0.23.1
